% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday Problem Functions}
\usage{
birthday(x)
}
\arguments{
\item{x}{Total number of people.}
}
\value{
Returns the chance that at least two people in x number of people have the same birthday.
}
\description{
Birthday Problem Functions
}
