% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Data Set}
\format{
## `fire`
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance in miles from fire station.}
  \item{DAMAGE}{Thousands of dollars}
  ...
}
}
\source{
<https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850>
}
\usage{
fire
}
\description{
A data set concerning damage from fire
}
\keyword{datasets}
