% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Uniform Distribution proving Central Limit Theorem}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{the number of samples in a distribution}

\item{iter}{the number of iterations}

\item{a}{the lower limit a sample can be}

\item{b}{the upper limit a sample can be}
}
\value{
a list
}
\description{
Uniform Distribution proving Central Limit Theorem
}
\examples{
w=myclt(n=50,iter=10000,a=5,b=10)

}
