% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Displays a shaded curve}
\arguments{
\item{mu}{integer that represents the mean of a distribution}

\item{sigma}{integer that represents the standard deviation of a distribution}

\item{a}{integer}
}
\value{
A shaded area between the curve and x axis from negative infinity to x = a.
}
\description{
Displays a shaded curve
}
\examples{
myncurve(mu = 10, sigma = 5, a = 6)

}
