% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Creates random samples and produces barplots}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{an integer that describes the size of each sample}

\item{iter}{an integer that describes the amount of times a trial should be run}

\item{time}{time between each iteration}
}
\description{
Creates random samples and produces barplots
}
\examples{
mysample(n=1000, iter=30, time = 1)
}
