% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Overbooking Problem}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of seats on the plane.}

\item{gamma}{The pain probability, aka the probability that we have more passangers show up than seats on the plane.}

\item{p}{The probability that a passenger shows up.}
}
\value{
The amount of seats that an airline should sell, based on the probability that a passenger shows up.
}
\description{
Overbooking Problem
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)

}
